requirements.txt
pillow
requests
fake-headers